cleaning snowbreds
