+++ src/tweet.py
@@ -1,0 +1,12 @@
+class Tweet:
+
+    def __init__(self, raw):
+        self.raw = raw
+        self.words = []
+        self.mapping = {}
+        self.emo_words = []
+        self.index = None
+        self.seed = None
+
+    def __str__(self):
+        return self.raw
